Update to API v2
- Library updated to support M2X API version 2
- Library file renamed to use a common versioning method. (m2x-MAJOR.MINOR.PATCH.js).
- M2XUVDemo updated to use API v2. M2XGPSDemo app is still using the previous version of the library (version 1) for illustration purposes.
- Documentation updated
- Cosmetic and typos fixed
