+++ demo/README.md
@@ -1,4 +1,4 @@
-These examples illustrates the following:
+These examples illustrate the following:
 
 * How to access one of the many sensors available on Gear S
 * How to leverage the Tizen M2X client to push data to M2X
@@ -6,4 +6,5 @@ These examples illustrates the following:
 * How to use Open Source libraries such as RequireJS, jQuery and CharJS to build a Gear S application
 
 ####IMPORTANT####
-You must replace the bogus M2X key in m2xclient.js with your own master key before building/running this demo. (search for 'new M2X') 
+* You must replace the bogus M2X key in m2xclient.js with your own master key before building/running this demo. (search for 'new M2X') 
+* For illustration purposes, M2XGPSDemo uses M2X V1 API library.
