+++ doc/api.js
@@ -2,10 +2,8 @@ YUI.add("yuidoc-meta", function(Y) {
    Y.YUIDoc = { meta: {
     "classes": [
         "M2X",
-        "M2X.batch",
-        "M2X.blueprint",
-        "M2X.datasource",
-        "M2X.feed",
+        "M2X.device",
+        "M2X.distribution",
         "M2X.key"
     ],
     "modules": [],
