+++ doc/classes/M2X.html
@@ -38,13 +38,9 @@
             
                 <li><a href="../classes/M2X.html">M2X</a></li>
             
-                <li><a href="../classes/M2X.batch.html">M2X.batch</a></li>
+                <li><a href="../classes/M2X.device.html">M2X.device</a></li>
             
-                <li><a href="../classes/M2X.blueprint.html">M2X.blueprint</a></li>
-            
-                <li><a href="../classes/M2X.datasource.html">M2X.datasource</a></li>
-            
-                <li><a href="../classes/M2X.feed.html">M2X.feed</a></li>
+                <li><a href="../classes/M2X.distribution.html">M2X.distribution</a></li>
             
                 <li><a href="../classes/M2X.key.html">M2X.key</a></li>
             
@@ -187,7 +183,6 @@
                     <div class="param-description">
                         <p>A object containing the following properties</br>
 <b>key</b>: M2X API key (String,required). Error will be thrown if key is not provided</br>
-<b>api_version</b>: M2X API version (Number(integer),optional). By default, will point to the latest stable M2X API version</br>
 <b>forward_errors</b>: (Boolean).If true, errors will be forwarded to the successful callback (if provided)</br></p>
 <p>Example:</p>
 <pre><code>var myM2X = new M2X({key:&#39;yourm2xkeygoeshere&#39;});</code></pre>
@@ -326,7 +321,7 @@ Callback signature</p>
             
         
         
-        <a href="../files/m2x.js.html#l1108"><code>m2x.js:1108</code></a>
+        <a href="../files/m2x.js.html#l1188"><code>m2x.js:1188</code></a>
         
         </p>
 
@@ -435,7 +430,7 @@ Callback signature</p>
             
         
         
-        <a href="../files/m2x.js.html#l1123"><code>m2x.js:1123</code></a>
+        <a href="../files/m2x.js.html#l1203"><code>m2x.js:1203</code></a>
         
         </p>
 
@@ -539,7 +534,7 @@ Callback signature</p>
             
         
         
-        <a href="../files/m2x.js.html#l1091"><code>m2x.js:1091</code></a>
+        <a href="../files/m2x.js.html#l1171"><code>m2x.js:1171</code></a>
         
         </p>
