+++ doc/classes/M2X.key.html
@@ -38,13 +38,9 @@
             
                 <li><a href="../classes/M2X.html">M2X</a></li>
             
-                <li><a href="../classes/M2X.batch.html">M2X.batch</a></li>
+                <li><a href="../classes/M2X.device.html">M2X.device</a></li>
             
-                <li><a href="../classes/M2X.blueprint.html">M2X.blueprint</a></li>
-            
-                <li><a href="../classes/M2X.datasource.html">M2X.datasource</a></li>
-            
-                <li><a href="../classes/M2X.feed.html">M2X.feed</a></li>
+                <li><a href="../classes/M2X.distribution.html">M2X.distribution</a></li>
             
                 <li><a href="../classes/M2X.key.html">M2X.key</a></li>
             
@@ -95,7 +91,7 @@
 
     
         <div class="foundat">
-            Defined in: <a href="../files/m2x.js.html#l980"><code>m2x.js:980</code></a>
+            Defined in: <a href="../files/m2x.js.html#l1060"><code>m2x.js:1060</code></a>
         </div>
     
 
@@ -249,7 +245,7 @@
             
         
         
-        <a href="../files/m2x.js.html#l1007"><code>m2x.js:1007</code></a>
+        <a href="../files/m2x.js.html#l1087"><code>m2x.js:1087</code></a>
         
         </p>
 
@@ -279,7 +275,7 @@
                     
 
                     <div class="param-description">
-                        <p>Key object. An error will be thrown if name(String) or permissions(Array) are not provided. See <a href="https://m2x.att.com/developer/documentation/keys#Create-Key">https://m2x.att.com/developer/documentation/keys#Create-Key</a> for details</p>
+                        <p>Key object. An error will be thrown if name(String) or permissions(Array) are not provided. See <a href="https://m2x.att.com/developer/documentation/v2/keys#Create-Key">https://m2x.att.com/developer/documentation/v2/keys#Create-Key</a> for details</p>
 
                     </div>
 
@@ -385,7 +381,7 @@
             
         
         
-        <a href="../files/m2x.js.html#l1024"><code>m2x.js:1024</code></a>
+        <a href="../files/m2x.js.html#l1104"><code>m2x.js:1104</code></a>
         
         </p>
 
@@ -521,7 +517,7 @@
             
         
         
-        <a href="../files/m2x.js.html#l1059"><code>m2x.js:1059</code></a>
+        <a href="../files/m2x.js.html#l1139"><code>m2x.js:1139</code></a>
         
         </p>
 
@@ -657,7 +653,7 @@
             
         
         
-        <a href="../files/m2x.js.html#l1074"><code>m2x.js:1074</code></a>
+        <a href="../files/m2x.js.html#l1154"><code>m2x.js:1154</code></a>
         
         </p>
 
@@ -746,7 +742,7 @@
             
                 <li class="arg">
                     
-                        <code class="optional">[feed]</code>
+                        <code class="optional">[deviceID]</code>
                     
                 </li>
             
@@ -793,7 +789,7 @@
             
         
         
-        <a href="../files/m2x.js.html#l986"><code>m2x.js:986</code></a>
+        <a href="../files/m2x.js.html#l1066"><code>m2x.js:1066</code></a>
         
         </p>
 
@@ -804,7 +800,7 @@
     </div>
 
     <div class="description">
-        <p>Retrieves a list of keys associated with the specified account.</p>
+        <p>Retrieves list of keys associated with the specified account.</p>
 
     </div>
 
@@ -816,7 +812,7 @@
             
                 <li class="param">
                     
-                        <code class="param-name optional">[feed]</code>
+                        <code class="param-name optional">[deviceID]</code>
                         <span class="type">String</span>
                         <span class="flag optional" title="This parameter is optional.">optional</span>
                     
@@ -824,7 +820,7 @@
                     
 
                     <div class="param-description">
-                        <p>It will list all the keys that are associated with that specific feed or its streams</p>
+                        <p>If provided, it will list all the keys that are associated with that specific device or its streams</p>
 
                     </div>
 
@@ -936,7 +932,7 @@
             
         
         
-        <a href="../files/m2x.js.html#l1039"><code>m2x.js:1039</code></a>
+        <a href="../files/m2x.js.html#l1119"><code>m2x.js:1119</code></a>
         
         </p>
 
@@ -947,7 +943,7 @@
     </div>
 
     <div class="description">
-        <p>Update name, stream, permissions, expiration date, origin or feed access of an existing key associated with the specified account</p>
+        <p>Update name, stream, permissions, expiration date, origin or device access of an existing key associated with the specified account</p>
 
     </div>
