+++ doc/index.html
@@ -38,13 +38,9 @@
             
                 <li><a href="./classes/M2X.html">M2X</a></li>
             
-                <li><a href="./classes/M2X.batch.html">M2X.batch</a></li>
+                <li><a href="./classes/M2X.device.html">M2X.device</a></li>
             
-                <li><a href="./classes/M2X.blueprint.html">M2X.blueprint</a></li>
-            
-                <li><a href="./classes/M2X.datasource.html">M2X.datasource</a></li>
-            
-                <li><a href="./classes/M2X.feed.html">M2X.feed</a></li>
+                <li><a href="./classes/M2X.distribution.html">M2X.distribution</a></li>
             
                 <li><a href="./classes/M2X.key.html">M2X.key</a></li>
